+++ user-login.php
@@ -118,7 +118,7 @@ include("_header.inc.php");
 <!--            <legend class="btn_toggle">Authentification</legend>-->
             <p>
                 <label for="pseudo" id="login_pseudo">Login</label>
-                <input type="text" name="pseudo" id="pseudo" value="<?= sanitizeForHtml($champs['pseudo']) ?>" size="30" />
+                <input type="text" name="pseudo" id="pseudo" value="<?= sanitizeForHtml($champs['pseudo']) ?>" size="30" autofocus />
 <!--                <div style="margin: 0 1em 1em 6em;font-size: 0.8em;line-height: 1.1em;padding: 0;"></div>-->
                 <?= $verif->getHtmlErreur("pseudo") ?>
 </p><p>